backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-vision
